tests/wpt/tests/css/css-grid/grid-definition/grid-shorthand-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: 'grid' shorthand does not reset gutter properties</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#grid-shorthand">
<meta name="flags" content="dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #grid {
    display: grid;
    grid-column-gap: 10px;
    grid-row-gap: 20px;
  }
</style>
<div id="grid"></div>
<script>
  var grid = document.getElementById("grid");

  test(
    () => {
      assert_equals(window.getComputedStyle(grid)["grid-template-columns"], "none");
      assert_equals(window.getComputedStyle(grid)["grid-template-rows"], "none");
      assert_equals(window.getComputedStyle(grid)["grid-column-gap"], "10px");
      assert_equals(window.getComputedStyle(grid)["grid-row-gap"], "20px");
    }, "Check gutter properties initial values");

  grid.style.grid = "100px / 200px";

  test(
    () => {
      assert_equals(window.getComputedStyle(grid)["grid-template-columns"], "200px");
      assert_equals(window.getComputedStyle(grid)["grid-template-rows"], "100px");
      assert_equals(window.getComputedStyle(grid)["grid-column-gap"], "10px");
      assert_equals(window.getComputedStyle(grid)["grid-row-gap"], "20px");
    }, "Check gutter properties after setting 'grid' shorthand");
</script>
